Module market_sol::lib::market::sol_market

SOL Market implementation specific to only that market and used across different modules.

Structs

SOLMarket

SOL Market struct, represents the market data with it’s goods, subscribers and metadata.

Constants

ALL_GOOD_KINDS 🔒

The array of all available goods for the market

LOCK_LIMIT 🔒

The maximum amount of locks for buy and sell operations. #default value = 10

MARKET_MARGIN 🔒

The margin this market applies on buy orders. #default value = 0.06

MARKET_NAME 🔒

The name of the market. #default value = SOL

TOKEN_DURATION 🔒

The validity period for the token given in days. #default value = 15

Functions

lock_limit_exceeded 🔒

Verifies if lock limit is still valid (below LOCK_LIMIT const)

log 🔒

Append log code to file according to specifications