Struct market_sol::lib::market::sol_market::SOLMarket

pub struct SOLMarket {
    pub(crate) goods: HashMap<GoodKind, Good>,
    pub(crate) subscribers: Vec<Box<dyn Notifiable>>,
    pub(crate) meta: MarketMeta,
}

SOL Market struct, represents the market data with it’s goods, subscribers and metadata.

Fields
goods: HashMap<GoodKind, Good>subscribers: Vec<Box<dyn Notifiable>>meta: MarketMeta
Implementations

impl SOLMarket

pub(crate) fn notify_everyone(&mut self, e: Event)

Notify every market including ours of an event

impl SOLMarket

pub(crate) fn new_with_quantities_and_path(
eur: f32,
yen: f32,
usd: f32,
yuan: f32,
optional_path: Option<&str>
) -> Rc<RefCell<SOLMarket>>

Returns SOLMarket with given quantites and optional file path to which the market will be serialized.

pub(crate) fn get_available_quantity(&self, good_kind: GoodKind) -> f32

Returns how much of the asked GoodKind is available (not locked)

fn get_exchange_rate(&self, good_kind: GoodKind) -> f32

Exchange rate (EUR/GoodKind) for a given good_kind

pub(crate) fn get_good_buy_exchange_rate(&self, good_kind: GoodKind) -> f32

Return the rate applied when the trader wants to BUY the good from this market The rate is EUR/goodkind

pub(crate) fn get_good_sell_exchange_rate(&self, good_kind: GoodKind) -> f32

Return the rate applied when the trader wants to SELL the good to this market The rate is EUR/goodkind

pub(crate) fn get_good_labels(&self) -> Vec<GoodLabel>

Returns vector of GoodLabel for each GoodKind availabale on the market

impl SOLMarket

This block contains the logic to serialize/deserialize the market to and from a file

pub(crate) fn write_to_file(&self)

If the market knows about a file, it means it read the state from there. This function updates such file with the current state of the market.

fn serialize_to_file_string(&self) -> String

pub(crate) fn read_quantities_from_file(path: &Path) -> Option<Vec<Good>>

Reads the file at the provided path and optionally returns vector with the goods that the SOL Market represented in that file has.

If there is an error reading or parsing the file, None is returned.