Function market_sol::showcase::test_overall_market_change_percentage

pub(crate) fn test_overall_market_change_percentage()

This function allows us to test how is the overall trend of our market, for any good. Ideally, we want this to be a slightly positive percentage.

Pro tip: if you care about execution speed, disable all prints since the I/O takes much longer than the computation